.history/resume_processing/requirements_20240822095435.txt
streamlit==1.16.0
openai==0.28.0
pytesseract==0.3.8
pdf2image==1.16.0
opencv-python==4.5.3.56
easyocr==1.5.0
langchain==0.0.200
pandas==1.3.3
python-dateutil==2.8.2
requests==2.25.1
torch==1.9.0
numpy==1.21.2
